requirements.txt
Pillow==10.0.1
Pillow==9.4.0
Pillow==10.3.0
streamlit==1.31.1
tensorflow==2.13.1
